requirements.txt
# Django>=3.0,<3.1
# wagtail>=2.8,<2.9

django
wagtail
django-recaptchapython-dotenv
